chore: remove ability to set visual failure threshold
According to our docs, the `FAILURE_THRESHOLD` and
`FAILURE_THRESHOLD_TYPE` environment variables were introduced when
running visual tests locally actually ran the suite on the host machine,
which could produce different results than in CI, as the test browser or
the fonts could not be the same.
Since then, we have changed how visual tests run locally. They now run
in a Docker container that's identical to the CI environment. This
doesn't always work as expected for different reasons, but makes it
irrelevant to have a way to override the failure thresholds. So this
removes the corresponding docs and code.