chore: cleanup the default MR template
Rewrites parts of the default MR template to reduce the noise incurred
by the checklist.
The acceptance checklist has been moved to a Markdown file in the
contribution docs, and we link to it from the MR template.
The integration MRs part has been made into a section within the
template.
A new section encourages authors to add screenshots or screen recording
when introducing visual changes.