feat(GlCollapsibleListbox): allow groups with no text
This allows groups of options without a name to be rendered at the top.
This can be useful when mixing general options (that require no label)
with specialized options (with a group label). `textSrOnly: true`
must be passed in the group. The `text` will still be used for
accessibility purposes.