fix(dropdowns): do not steal focus back from the consumer
When a dropdown gets closed after the user selected an option, the
consumer might want to move the focus to some other element (a form
input for example). In such case, dropdowns should not attempt to
re-focus their toggle as that might interfere with the consumer's logic.

This ensures that the `closeAndFocus` method gets aborted if it detects
that the focus has been taken over.