fix(GlBreadcrumb): Fix details in resize calc
For the first size check with all items visible, it is not needed to
consider the collapse button width, as it wouldn't be shown in case all
items fit.

Also shrink the expected size of the button to not expect more space
than what is actually required. This otherwise can lead to recursive
calls, shrinking down to the last element because the actual wrapper
size is smaller than what the calc expects.