feat(BaseDropdown): Use a "main" element boundary when available
BREAKING CHANGE: This change slightly alters the default positioning
of GlDisclosureDropdown and GlListbox.

If the dropdown is placed inside a <main> element: The disclosure
will use main as a boundary to not overflow it. This change aims
to smartly reduce the need to use custom "placement" ('right'
and 'bottom-end') options, especially when the dropdown is placed
on the right side, such as in contextual menus.

If the dropdown is not placed in a <main> element, behavior does
change.