fix(dropdowns): ignore vertical boundaries
When dropdowns are wrapped within a `<main>` wrapper, we want them to
consider it the wrapper only for the left and right sides. Vertically,
the dropdowns should consider the document as their boundary element.
This prevents issues where a short `<main>` wrapper would cause dropdown
to be top-positioned even though they have room to be bottom-positioned.