refactor(OutsideDirective): Remove event timestamp checking code
The GlOutsideDirective now uses event capturing, which means that its
timestamp-checking behaviour can be removed.

This behavior was added before to prevent calls on the event listener
before the component with the outside click directive was actually
rendered into the DOM.

By adding E2E tests with Cypress for this behavior, we can delete the
Unit tests written in Jest which tested the timestamp-checking
behaviour.