ci: Disable publish-to-npm-dry-run in fork pipelines
semantic-release does actually use the remote branches for
double-checking it's config:
https://github.com/semantic-release/semantic-release/blob/v19.0.5/lib/branches/index.js

Unfortunately when we manually run a merge pipeline on a fork MR, the
pipeline runs in the context of the upstream project, where the fork
branch doesn't exist, leading to an error.

The easiest fix is to just disable the dry-run job for these kind of
MRs. (Alternatively we could change the remote in the job, but that
sounds messy?)