feat: Add vendored bootstrap@4.6.2 SCSS
@gitlab/ui now ships it's own vendored version of bootstrap SCSS. In a
future version of GitLab UI we will stop relying on the bootstrap peer
dependency. This will then be marked as a breaking change. In order to
prepare for the breaking change, please:

1. Make sure your locally installed bootstrap version matches 4.6.2
2. Scan your CSS / SCSS imports for imports of `bootstrap/scss/`. Adjust
   them to `@gitlab/ui/scr/vendor/bootstrap/scss`. As the vendored
   version exactly matches bootstrap@4.6.2, this should not result in
   any difference of your compiled CSS.
3. This is all which should prepare you for the breaking change.

Once the breaking change lands: Unless you use the JavaScript bits of
bootstrap, you can simply drop the bootstrap dependency.