docs(css): updates our stance with regard to CSS utils
As we are replacing our custom CSS utils library with Tailwind CSS, it's
time for a docs update and cleanup.

This documents how consumers should setup Tailwind CSS on their end, and
what stylesheets they need to import.

This also updates and/or removes a bunch of docs related to our
historical architecture decisions.