ci: create integration branches in a fork
The main GitLab project has grown too big and pushing integration
branches to it over HTTPS was timing out most of the time. We are
switching to a fresh fork which will only be used for integration
branches and should therefore remain small enough that push operations
won't fail any time soon.