chore: make tailwind less of a headwind
Hi Mark! Thank you for nerd-sniping me into this one. This was a wild
ride of trying to debug what's going on, especially considering the fact
that it triggered webpack PTSD with relation to CSS imports and such.

So suddenly storybook takes minutes to start instead of seconds.
It must be something like this performance regression from storybook@7,
where upgrading to storybook 8 will really help
https://github.com/storybookjs/storybook/issues/26517. Only logical
conclusion, let's try to upgrade to storybook@8 which we cannot really
do because they dropped support for Vue@2 like any sane project would.

But hey, we have an environment level flag, to use Vue@3. So why not
upgrade partially, and try the Vue@3 version. 60 minutes into upgrading
storybook and reading their deprecations and whatnot, boom. Build fails.
But before failing, it just takes 3 minutes start.

Okay, at this point we are in shambles and just suspect that it might be
webpacks fault. How could it not be? By the way, why are webpack dev
Tobias Koppers and rollup dev Lukas Taegert-Atkinson both German? Do
Germans have a natural proclivity to immerse themselves in build tooling
for the blessed JS ecosystem? Is this the reason I am here doing this
instead of adding fields to GraphQL entities?

If even the Germans and the predisposition to satisfy their need for
self-loathing and masochism through the remedy that is build tools is
not helping, let's go French, at least by name, and our fast salvation
will be vite. I mean, swapping out webpack, SWC and esbuild for just
vite and esbuild must solve this, right?

After getting it to work, storybook starts in just 10 seconds. Sweet
victory at last after two hours in. Let's whip up storybook in the
browser. Blank pages for three minutes. Back the drawing board.

In this moment it should have clicked because it took some CSS assets
for three minutes to load. But instead of looking into what we changed
recently with regards to CSS, I revert back to webpack and start looking
into the CSS tool chain we have set up for webpack. Faint memories and
PTSD of OOM issues with relationship to sass-loader and css-loader and
mini-css-extract-plugin and cache-loader start to re-emerge. After
trying at least 10 permutations of loaders, disabling source maps,
moving to sass-embedded, the problematic POS (piece of software) is
finally cornered: it must be tailwind.

But how? Tailwind performance is not a problem in the GitLab project
which is much bigger, _why_ would it be so problematic in a much smaller
project. After some assumptions about webpacks importing logic, trying
to run tailwind directly in the project takes minutes. But why? We scan
the gitlab-ui source in the GitLab project as well.

Because we scan everything in the src folder. All of bootstrap-vue, it's
node_modules, all the specs files, all the story files.

What a ride. Three hours I am never going to get back. Apart from that
no moral to this story, just a Grimm tale which I won't tell my kids.