chore: Get rid of not one but two C compilers
Apparently the `cobertura-merge` dependency had some transitive
dependencies (`nan` and `node-expat`) which compiled some C extensions
for node. The dependency was introduced in:
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/4301

The problem ensued because some of the playwright  and cypress images
shipped slightly different node versions than what we are caching
dependencies with. Subsequently we added two C compilers:
- https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/4452
- https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/4520

Seems like we are not the only ones hitting the problem and someone
created `merge-cobertura`: https://www.npmjs.com/package/merge-cobertura
which doesn't rely on XML parsers with native extensions. This one is
likely a little slower. But given that this runs in CI, probably not a
big problem.