chore(deps): Update @vue/test-utils@2
This new version includes
https://github.com/vuejs/test-utils/commit/74c9af4e35430fb0c3b5cd7366d349b1a8160fd9,
which means we need to differentiate between Vue warnings and Vue
errors.

We also add `@vue/server-renderer` as a dev dependency. It *should* be
marked as an optional peer dependency, but there hasn't been a release
of VTU@2 with the fix included yet:
https://github.com/vuejs/test-utils/pull/2511.

Furthermore, even though it's "optional", if we don't install it, Jest
complains about not being able to find it, since it ends up being
`require`d in the top level scope of the VTU@2 module.

Finally, the new version also includes a fix that means we can remove
our warning ignore list: https://github.com/vuejs/test-utils/pull/2413.