chore: Fix toast visual test in Vue 3
It seems that in Vue 2, toasts are automatically removed when the parent
component is destroyed. In Vue 3, that appears not to happen, for some
reason.

Ideally that underlying problem would be fixed, but in practice, it
shouldn't be much of a problem, since toasts auto-hide eventually
anyway.

The reason this matters in the visual tests is that we switch stories
between taking screenshots, which means that toast(s) from the previous
story might still be present. This meant that the `getByRole('status')`
query of the `play` function was failing, since more than one toast was
present sometimes.

Now, we explicitly hide toasts when the stories are destroyed.