chore(DesignTokens): Improve performance of TokensTable story
The TokensTable story was rendering all tokens at once, which is over
1,000 rows. This had quite poor performance, and made interacting with
the table unpleasant. This was worsened by the lack of debounce on the
search field.

This commit improves performance by:

- debouncing on the search field, so re-renders aren't triggered on
  every key press;
- paginating tokens, so they're not all rendered at once.

It also improves the searching behaviour by using fuse.js to do fuzzy
searching, rather than the basic filtering provided by GlTable:

- searches are fuzzy, so exact substring matches are no longer required;
- results are sorted by closeness to the query (see
  https://www.fusejs.io/concepts/scoring-theory.html);
- only the name, value, description and deprecated fields are now
  searched, previously all were (see
  https://bootstrap-vue.org/docs/components/table#filtering);
- the standard GlSearchBoxByType component is used for the search input,
  providing a search icon and clear button.

Finally, this switches to a fixed table layout, which is generally
faster for the browser to render, and makes sure the value column text
wraps if needed.