fix(GlBreadcrumb): Hide breadcrumb item if it has no text
This fixes an accessibility problem that can occur in GitLab where a
link is rendered for a breadcrumb item that has no text. The text and
href are being removed from that breadcrumb item so that it won't show,
but the element is still visible to screen readers. Insead, we should
make the item hidden if it has no text.