chore: Unify & fix default prop values in stories
This commit extracts and uses a `propDefaultFactory` story helper across
many slightly buggy stories.

Many stories defined a `defaultValue`/`propDefault` helper that returned
the default value of the given prop. They all returned the incorrect
value in the case that the default value was a function.

When the default value is a function, its return value is the default
value, not the function itself.

The newly extracted helper handles this almost correctly. Almost,
because Vue 2/3 pass the Vue instance/raw props as a first argument.
This helper does not, because it doesn't have access to them.