Follow GitLab's commit message guidelines
We were using semantic commits for updating the changelog, and tagging
and publishing GitLab releases. Since
https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/4075,
we no longer update the changelog, tag or publish releases. So, the need
for using semantic commits has gone.

For simplicity and consistency, we now only require that commit messages
conform to GitLab's own commit message guidelines, via the
[Danger Review CI component](https://gitlab.com/gitlab-org/components/danger-review).

Finally, semantic commits are not suitable for a monorepo, which is what
we intend on turning this project into. See
https://gitlab.com/gitlab-com/gitlab-OKRs/-/work_items/10538 for the
details.