Add publish job and script
With this, the @gitlab/release-tooling-test package can in theory be
version bumped, published, committed, tagged and pushed on the default
branch.

It's worth discussing the hack in this commit.

Changesets can be used in either monorepos or single-package
repositories. It distinguishes between these by the use of workspaces,
which most npm-compatible package managers support in some form.

If changesets does not detect the use of workspaces, then it assumes all
package releases are for the root package, i.e., design.gitlab.com.

That is not what we want. So, the logical next step would be to set up
workspaces for this project. But we haven't, because we want to
eventually make GitLab UI a package within this project, and the
particular {d,devD,peerD}ependencies of design.gitlab.com and GitLab UI
make yarn@1.x throw a nonsense error when we try:

    warning Workspaces can only be enabled in private projects. error An
    unexpected error occurred: "expected workspace package to exist for
    \"webpack\"".

The first line is nonsense because the root package.json already has
`private: true` set.

The second line is nonsense because it has no reason to expect there to
be a workspace package for webpack.

To fix this properly, we will need to move to a modern, supported
package manager. All the latest versions of npm, yarn and pnpm correctly
install this project with GitLab UI as a workspace package.

Once we've done that, we can adopt workspaces properly and remove this
hack.