Disable MD059 Descriptive Link Text Linting
markdownlint has a new lint:
https://github.com/DavidAnson/markdownlint/blob/main/doc/md059.md

This is supposed to prevent lint against markdown links which are like:
`[more]` or `[link]`. While a good idea, the lints on design.gitlab.com
were like `[link] component` (which is in line with `[button]
component`) or explicitly calling out that `[more]` should be `[Learn
more]`.

So this linting rule doesn't really apply to our code base, let's
disable it.