Multi deploy all the pages
This creates three additional pages environments per pipeline:

1. One for storybook: `/storybook` on main, `/review-mr-<iid>-storybook`
   for MRs
2. One for storybook@vue3: `/storybook-vue3` on main,
   `/review-mr-<iid>-storybook-vue3` for MRs
3. One for tailwind config viewer: `/tailwind` on main,
   `/review-mr-<iid>-tailwind` for MRs

The existing review app for design.gitlab.com is also tied to the the
new pages job for `storybook`, so that the iframes are shown from there.

Definitely was a one-shot vibe code thing and we did not need multiple
attempts because CI is super easy.