Fix pipeline creation
In https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/4494#note_2528118101,
the `publish_npm_packages` job initially had `needs` defined, to make the
relationship between jobs explicit. I suggested
this wasn't maintainable, and to instead use `dependencies`.

That doesn't work, as it seems that even when `publish_npm_packages`
won't be created due to its `rules` (e.g., it's a merge request
pipeline), its `dependencies` are still inpsected, and if they're found
not to exist on the pipeline, it fails. At least, that's what appears to
be happening.

So, this reverts to the previous approach: relying purely on the `stage`
to determine when the `publish_npm_packages` job runs.

An alternative approach would be to use `needs`, but only add jobs which
we know are real dependencies, i.e., `ui:build_package`, for artifacts,
and `production`, to not interfere with Pages deploys. Everything else
could be assumed to be successful, via merge train pipelines.