Create releases for published packages
This makes the `publish_npm_packages` job create releases for published
npm packages.

The main benefit of this is that it would allow
[Renovate](https://gitlab.com/gitlab-org/frontend/renovate-gitlab-bot)
to display the relevant changelog in dependency update merge requests.
For example:
- a recent update MR _without_ a changelog section:
  https://gitlab.com/gitlab-org/gitlab-svgs/-/merge_requests/1179
- an older update MR _with_ a changelog section:
  https://gitlab.com/gitlab-org/gitlab-svgs/-/merge_requests/1173

There are improvements we could make in future. For example, adding
links to the commits in the changelog. I've not done that here to keep
this small. I'm also unsure whether Renovate will
[detect](https://docs.renovatebot.com/key-concepts/changelogs/) the
releases correctly, and if it does, how it will render them. So, my plan
is to get basic releases implemented, and iterate.

Another improvement could be to handle dependent releases (where
packages are bumped when other packages they depend on are bumped). For
now, we only have one package, so this doesn't matter. In future, when
we have more packages, it's not essential for the release description to
have this information. The changelog files themselves will have this
information.

The core logic of building a changelog entry from available changesets
is based on @changesets/apply-release-plan@7.0.12:
https://github.com/changesets/changesets/blob/baa9d45384213d7a697873848dbeb183001fd243/packages/apply-release-plan/src/get-changelog-entry.ts#L52-L59