Fix ui:update_screenshots job
Previously, the push was failing because the `GITLAB_TOKEN` CI variable
is [protected][1], so it's not available in unprotected branches/tags,
i.e., merge request pipelines.

The old GitLab UI project did not protect that variable, which is why it
worked there. That's not great, as that token has full `api` scope, and
so doesn't follow the principle of least privilege.

Here, we are using a new project access token, with scope limited to
`write_repository`, and is not a protected CI/CD variable.

[1]: https://docs.gitlab.com/ci/variables/#protect-a-cicd-variable