GlLineChart: Add tooltip click-to-pin feature
Adds a new prop ('clickToPinTooltip'), which when
set to true will allow the chart's tooltip to be
pinned to the current position by clicking
on the chart.

This allows the user to interact with the
tooltip's content, for example click links.

A second click on the chart, or pressing the
'escape' key, will unpin the tooltip.