Simplify truncate text spec
This is needed for the future migration to Jest 30.

The latest version of jsdom (or rather cssstyle, which implements this
part of jsdom) no longer seems to support whatever iteration method
`toContain` is using.

Anyway, checking the property value is sufficient. There's no need to
check that the property exists first.