Fix isVisible test under Jest 30
Without mounting the element into the actual DOM, the
`click on action item executes its function` test will fail under Vue 3.
Presumably, this is due to a stricter treatment of disconnected elements
passed to the `getComputedStyle` implementation in the newer jsdom.

It's not clear why other tests passed, but either way, this fixes this
one.