Hoist prettier configuration files
The trailingComma option default was changed to "all" in v3.0.0, so
there's no need to specify it now. See
https://prettier.io/docs/options#trailing-commas.

The migrate_custom_utils_to_tw.bundled.mjs file was added to the ignore
file as well, since, if it exists, it can take a few seconds to process
on its own, and it's a generated file anyway.