Serve Nuxt site from localhost only
This ensures that Nuxt is not available on the local or public network
by default.

This is similar to 35cd39e0e8bd6000b91802787c262965d4b9aa31.

To make it available on the network (e.g., to test on a different
device) run:

    HOST=<your IP> yarn start

or

    HOST=0 yarn start

to serve it from all network interfaces (not recommended).