Add basic live example editor
This adds a minimally viable live example editor, via the
[`vue-live`](https://github.com/vue-styleguidist/vue-live/tree/v1.17.2)
package. It was chosen because it's light weight, and built in and for
Vue.

Known limitations:

- Editor syntax highlighting is not Dark Mode compatible. This is
  affected by various CSS sources, like Bootstrap itself, and our own
  framework CSS.
- The editor has no code formatting capabilities, like auto-indentation,
  or prettifying.
- The renderer is in-document, meaning that media queries wouldn't apply
  to it if we wanted to implement a viewport resizing feature. To fix
  that, we'd need to render into an `iframe`.
- By being in-document, there is some CSS leakage from Nuxt affecting
  the rendering of some components. For example,
  `<gl-button href="#">foo</gl-button>` has underlined text.

If we decide we need more functionality in future, we can swap
`vue-live` out for something else. Other possibilities include:

- [@vue/repl](https://github.com/vuejs/repl)
  - Vue 3 only
  - Would need to to implement our own playground based on this, like
    https://github.com/vuejs/core/tree/v3.5.17/packages-private/sfc-playground
- [Sandpack](https://sandpack.codesandbox.io/)
  - Heavy weight
  - Would need to implement our own `sandpack-vue` component, or
    embed/wrap `sandpack-react`.

Both of those are heavier, require a much larger implementation effort
on our side, and might be more powerful than we would need anyway.

Addresses https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/issues/2872.