fix(GlBreadcrumbs): Add medium variant of collapse calculation
Even thought the dropdown toggle of collapsed breadcrumbs has the
same size in small and medium breadcrumbs, the breadcrumb item that
wraps it is slightly wider on the medium size. This results in a
cycle of collapsing and uncollapsing the first item, when it has a
certain width. An extra condition to change buttonWidth value
based on the component size, will allow the breadcrumbs to collapse
only when necessary.