Fix alert page search box
In production only, if you visit the `/components/alert` page directly,
the sidebar's search box doesn't render.

In the console, there's this error:

```
DOMException: Node.appendChild: Cannot add children to a Text
  > NuxtJS
```

This fix is _bizarre_.

Here are a couple of patches that fix the problem:

- Remove HTML syntax highlighting on a code block.
   ```diff
   diff --git c/contents/components/alert.md i/contents/components/alert.md
   index f16c011b90..8aa2c65c02 100644
   --- c/contents/components/alert.md
   +++ i/contents/components/alert.md
   @@ -117,7 +117,7 @@ Alerts don't handle their own visibility, so it's the parent component's
    responsbility to listen for the `dismiss` event and hide the alert in some way.
    For example:

   -```html
   +```
    <script>
      ...
      computed: {

   ```
- Don't do property access at the beginning of an expression?
   ```diff
   diff --git c/contents/components/alert.md w/contents/components/alert.md
   index 0697bc33f9..86155ad257 100644
   --- c/contents/components/alert.md
   +++ w/contents/components/alert.md
   @@ -124,7 +124,7 @@ For example:
      ...
      computed: {
        shouldShowAlert() {
   -      return !this.isAlertDismissed && this.someOtherCondition();
   +      return !isAlertDismissed && this.someOtherCondition();
        },
      },
      ...
   ```

It appears that a minimal reproduction is to have a page with this in it:

````
```html
<script>a.b</script>
```
````

Figuring out _why_ that breaks anything at all probably means digging
into the guts of Nuxt. For now, we'll work around it.

Addresses https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/issues/2953.