Fix sort order of tokens to respect alphanumerical order
Currently the `item` prop provided to `<gl-table>` is sorting the items
alphanumerically after filtering deprecated items to the end.
The issue with this is that items display in an incorrect order
e.g. something-1, something-11, something-2, something-25 etc.

Update to maintain the behaviour to sort deprecated items to the end.