Differentiate deprecation copy for components and other patterns
Currently our deprecation warning says "Please refrain from using this
component - it is deprecated and will be removed." This copy is a bit
awkward/not relevant when the docs page in question is a pattern
documentation page.

This MR introduces differentiated deprecation copy so that components
will say "component" and any other page will say "pattern".