Unify node images across CI
Rather than using Alpine here and Debian there, use the same image for
all node jobs.

As the monorepo grows and build interdependencies between packages
increase, having a consistent base image to work with will simplify
things.

A couple of CI failures prompted this, e.g., from the
`svgs:build-and-check` job:

    style-dictionary@5.0.4: The engine "node" is incompatible with this
    module. Expected version ">=22.0.0". Got "20.19.5"

and from the `ui:visual` job:

    ./build_scripts/build-svgs.sh: line 31: unzip: command not found

We could have added `apk add unzip` to the `.node-job` definition in the
root CI file, but the reason for that would be distant: it's only needed
because the prepare script runs a `@gitlab/svgs` build step.