Only update peer dependencies when out of range
This enables the experimental Changesets option
`onlyUpdatePeerDependentsWhenOutOfRange`. From the [docs]:

> When set to true, Changesets will only bump peer dependents when
> peerDependencies are leaving the range.

Without this change, bumping `@gitlab/svgs` with even a patch version
would cause `@gitlab/ui`'s peer dependency on `@gitlab/svgs` to be
updated (tightened) unnecessarily, and for `@gitlab/ui` to get a *major*
bump.

That is entirely undesirable and unnecessary. I don't know why this
experimental option defaults to `false`, when `true` is a more sensible
default.

Because of the experimental nature of this option, which could be
removed in even patch versions of Changesets, I've reduced the set of
reviewers assigned to dependency updates of it in
https://gitlab.com/gitlab-org/frontend/renovate-gitlab-bot/-/merge_requests/1286.

[docs]: https://github.com/changesets/changesets/blob/main/docs/experimental-options.md